fix: <ShopInformationColumn /> としてコンポーネントに分割
